refactor: extract ttymap-core (event / command / config / keymap) (#351 Phase A)
First mechanical move of the workspace split planned in #351:
factor out the cross-cutting vocabularies that would otherwise force
a circular dependency between the upcoming `ttymap-tui` (UI
primitives) and `ttymap-lua` (plugin runtime) crates.

New `ttymap-core` crate holds:

- `UserCommand` — the GoF Command vocabulary every emission site
  (palette, plugin callbacks, mouse adapter, …) speaks.
- `EventBus` / `Event` / `Level` — Lua-agnostic pub/sub primitive.
- `Config` / `RuntimeConfig` — settings shape mutated by Lua's
  `ttymap.opt.*` and consumed by every layer.
- `KeyMap` / `KeybindingOverrides` — `key → UserCommand` resolution
  table. Lives here (not in `input/`) because it names both the
  Command vocabulary and the user-configurable override shape.

Depends on `ttymap-engine` only (LonLat / ThemeId / MapAction /
engine::Config). No `ratatui` / `mlua`.

`ttymap-app` consumes the moved types via `pub use ttymap_core::*;`
re-exports, so every existing `crate::event::*` / `crate::command::*`
/ `crate::config::*` / `crate::input::keymap::*` call site keeps
resolving without churn. Phase B (extract `ttymap-tui`) will collapse
the re-exports as part of moving `input/` out of `ttymap-app`.
