+++ ttymap-core/src/command.rs
@@ -40,9 +40,9 @@
 //! off `UserCommand` means the focus state machine isn't coupled to
 //! the dispatch table.
 
-use crate::input::KeyMap;
-use crate::theme::ThemeId;
+use crate::keymap::KeyMap;
 use ttymap_engine::map::MapAction;
+use ttymap_engine::theme::ThemeId;
 
 /// What the app can do in response to an event. Emitted by palette
 /// providers, plugin handlers, and async plugin polling; interpreted
