+++ ttymap-core/src/event/bus.rs
