+++ ttymap-core/src/event/payload.rs
